refactor(indexes): Split indexes.py in multiple files
